Move account functions to remove import-outside-toplevel
This is part of the work to remove cyclic imports from the backend
