+++ backend/infrahub/core/__init__.py
@@ -17,8 +17,6 @@ if TYPE_CHECKING:
     from infrahub.graphql.types import InfrahubObject
     from infrahub.types import InfrahubDataType
 
-# pylint: disable=redefined-builtin
-
 
 def get_branch_from_registry(branch: Optional[Union[Branch, str]] = None) -> Branch:
     """Return a branch object from the registry based on its name.
@@ -86,57 +84,6 @@ async def get_branch(branch: Optional[Union[Branch, str]], session: Optional[Asy
     registry.branch[branch] = obj
 
     return obj
-
-
-async def get_account(
-    account,
-    session: AsyncSession,
-    branch=None,
-    at=None,
-):
-    # No default value supported for now
-    if not account:
-        return None
-
-    if hasattr(account, "schema") and account.schema.kind == "Account":
-        return account
-
-    # Try to get it from the registry
-    #   if not present in the registry, get it from the database directly
-    #   and update the registry
-    if account in registry.account:
-        return registry.account[account]
-
-    # pylint: disable=import-outside-toplevel
-    from infrahub.core.manager import NodeManager
-
-    account_schema = registry.get_schema(name="Account")
-
-    obj = await NodeManager.query(
-        account_schema, filters={account_schema.default_filter: account}, branch=branch, at=at, session=session
-    )
-    registry.account[account] = obj
-
-    return obj
-
-
-def get_account_by_id(id: str):  # pylint: disable=unused-argument
-    # No default value supported for now
-    # if not id:
-    return None
-
-    # from .account import Account
-
-    # if id in registry.account_id:
-    #     return registry.account_id[id]
-
-    # obj = Account.get(id=id)
-    # if not obj:
-    #     return None
-
-    # registry.account[obj.name.value] = obj
-    # registry.account_id[id] = obj
-    # return obj
 
 
 @dataclass
